+++ src/epics/apiEpics.js
@@ -3,6 +3,7 @@ import { combineEpics } from 'redux-observable';
 
 import {FETCH_INITIAL_NODE, RECEIVE_INITIAL_NODE, receiveInitialNode} from '../actions/apiActions';
 import {REQUEST_SINKS, getSinks, REQUEST_SOURCES, getSources} from '../actions/dragonflyActions';
+import {ADD_TRIPLET, addEdge} from '../actions/d3actions';
 import {addNode} from '../actions/d3actions';
 import rootApi from '../api/rootAPI';
 
@@ -30,11 +31,40 @@ const fetchSources = action$ =>
         .mergeMap(({node}) => Rx.Observable.from(API.getAllParents(node.hash)))
         .map(getSources)
 
+/**
+ * Fires when a triplet is added.
+ * Because we want to add edges between
+ * the new triplet and any nodes already
+ * placed, grab all sources and sinks for
+ * the subject and predicate, then fire
+ * ADD_EDGE for each triplet.
+ */
+const onAddTripletAddEdge = action$ =>
+    action$.ofType(ADD_TRIPLET)
+        .mergeMap(({triplet}) =>
+            Rx.Observable.merge([Rx.Observable.from(API.getAllParents(triplet.subject.hash))
+                                    .flatMap(v => v)
+                                    .map(v => addEdge(v, v.predicate, triplet.subject)),
+                                Rx.Observable.from(API.getAllChildren(triplet.subject.hash))
+                                    .flatMap(v => v)
+                                    .map(v => addEdge(triplet.subject, v.predicate, v)),
+                                Rx.Observable.from(API.getAllParents(triplet.object.hash))
+                                    .flatMap(v => v)
+                                    .map(v => addEdge(v, v.predicate, triplet.object)),
+                                Rx.Observable.from(API.getAllChildren(triplet.object.hash))
+                                    .flatMap(v => v)
+                                    .map(v => addEdge(triplet.object, v.predicate, v)),]))
+        .flatMap(v => v)
+
+        
+        
+
 const apiEpics = combineEpics(
     fetchNode,
     receiveNode,
     fetchSinks,
-    fetchSources
+    fetchSources,
+    onAddTripletAddEdge
 )
 
 export default apiEpics;