+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+ - Add edges between visible nodes
  - Recenter graph
  - Zoom and panning
  - Dragonfly add triplet